+++ packages/sources/src/index.ts
@@ -18,6 +18,8 @@ export * from './components/button/button.models';
 export * from './components/button/button.stories';
 export * from './components/conclusion/conclusion.models';
 export * from './components/conclusion/conclusion.stories';
+export * from './components/context/context.models';
+export * from './components/context/context.stories';
 export * from './components/date-picker/date-picker.models';
 export * from './components/date-picker/date-picker.stories';
 export * from './components/definition-list/definition-list.models';